pyproject.toml
[tool.poetry]
name = "desctop-server"
version = "0.1.0"
description = ""
authors = ["alploskov <ploskov-ai@yandex.ru>"]
readme = "README.md"
packages = [{include = "desctop_server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
